+++ classes/class.xoctLog.php
@@ -46,7 +46,12 @@ class xoctLog extends ilLog {
 	 */
 	public static function getInstance() {
 		if (! isset(self::$instance)) {
-			self::$instance = new self(ILIAS_LOG_DIR, self::OD_LOG);
+			if (ILIAS_LOG_DIR === "php:/" && ILIAS_LOG_FILE === "stdout") {
+                // Fix Docker-ILIAS log
+                self::$instance = new self(ILIAS_LOG_DIR, ILIAS_LOG_FILE);
+            } else {
+				self::$instance = new self(ILIAS_LOG_DIR, self::OD_LOG);
+			}
 		}
 
 		return self::$instance;
@@ -85,7 +90,12 @@ class xoctLog extends ilLog {
 	 * @return string
 	 */
 	public function getLogFile() {
-		return self::OD_LOG;
+			if (ILIAS_LOG_DIR === "php:/" && ILIAS_LOG_FILE === "stdout") {
+                // Fix Docker-ILIAS log
+               return ILIAS_LOG_FILE;
+            } else {
+				return self::OD_LOG;
+			}
 	}